Add edit course functionality
Implemented edit course functionality which allows users to select a course and update its details. As part of this change, added a new 'loadById' method in the 'courses.service.ts' to fetch a course by id. Also added a course resolver to aid in loading the course before navigating to the edit page, ensuring the form is populated with current course details. Removed an unnecessary null option from category select list.